+++ src/@types/table.d.ts
@@ -1,9 +1,17 @@
 import { RowData } from '@tanstack/react-table';
 
-import { ColumnMeta as ColumnMetaOptions } from '../types';
+import { ColumnMeta as ColumnMetaOptions, TablePaginationConfig } from '../types';
 
 declare module '@tanstack/react-table' {
   interface ColumnMeta<TData extends RowData, TValue> extends ColumnMetaOptions {}
+
+  interface TableMeta<TData extends RowData> {
+    /**
+     * Pagination settings used when applying query column filters.
+     * @type {TablePaginationConfig}
+     */
+    pagination?: TablePaginationConfig;
+  }
 
   interface FilterMeta {
     from?: number;